New `fchdir` function
According to the man page chroot(2):

	This call does not change the current working directory, so that
	after the call '.' can be outside the tree rooted at '/'. In
	particular, the superuser can escape from a "chroot jail" by
	doing:

		mkdir foo; chroot foo; cd ..

The rpm library enters to the rootfs by chdir("/") to root and by
chroot("/path/to/rootfs") to rootfs then[1].

It exits from the chroot environment by chroot(".") to cwd (ie. /) and
fchdir("/path/to/rootfs") (ie. /path/to/chroot) latter[2].

In fine, the rpm library makes sure that the process can setup a chroot
environnement and can escape from it then; the process root directory
(ie. /path/to/rootfs) is a subdirectory of the cwd directory (ie. /).

This implements the fchdir(3) function that is used by the rpm library
but does not implement the escaping from the chroot jail if chrooting
outside.

TODO: Implement the feature correctly (i.e. path is outside the chroot)
and check for superuser privileges.

[1]: https://github.com/rpm-software-management/rpm/blob/257077a60c719bf98915c17fa71dc6c68fb0c3e9/lib/rpmchroot.c#L129
[2]: https://github.com/rpm-software-management/rpm/blob/8ed452dd8697917637d9e6fcc4b8c5d4bb50c8d0/lib/rpmchroot.c#L155