New variable FAKECHROOT_EXCLUDE_SYMLINK is introduced
The fakechroot command expands all absolute symlinks with the chroot
base directory.

But fakechroot is also used to create a root file-system and it is
required to keep these absolute symlinks not expanded; the symlinks will
be broken once the rootfs is mounted.

Some utilities make absolute symlinks that should not be expanded as
they are local to the rootfs.

For example, systemctl creates absolute symlinks to enable new services
below the directory /etc/systemd. The machine-id is also a symlink.