New variable FAKECHROOT_EXCLUDE_SYMLINK is introduced
TL;DR; This new variable FAKECHROOT_EXCLUDE_SYMLINK is analog to the
variable FAKECHROOT_EXCLUDE_PATH but for the symlinks, more precisely
the symlink values. The symlink values that are absolute path *AND* that
are created by either symlink(3p) or symlinkat(3p) *AND* that are listed
in the variable FAKECHROOT_EXCLUDE_SYMLINK are *NOT* expanded with the
chroot base directory: the value remain local to the chroot. However,
the canonical file name path is still expanded.

Here is an example:

	$ FAKECHROOT_EXCLUDE_SYMLINK=/absolute/path/in/chroot/value
	$ export FAKECHROOT_EXCLUDE_SYMLINK

	$ chroot /path/to/chroot /bin/sh
	# (...)
	# ln -s /absolute/path/in/chroot/value any-path-in-chroot-filename
	# readlink path/in/chroot/filename
	/absolute/path/in/chroot/value
	# exit

	$ readlink /path/to/chroot/any-path-in-chroot-filename
	/absolute/path/in/chroot/value
	$ file /path/to/chroot/any-path-in-chroot-filename
	/path/to/chroot/any-path-in-chroot-filename: broken symbolic link to /absolute/path/in/chroot/value

The fakechroot command expands all absolute symlink values with the
chroot base directory once they are created via either symlink(3p) or
symlinkat(3p).

But fakechroot may also be used to create a root file-system and it is
required to keep these values local to the chroot environment (i.e. not
expanded with the chroot base directory).

The symlink values are accessible within the chroot environment but are
broken outside (of refer to the host node) if the chroot environment is
not set.

Some utilities make absolute symlink values that should not be expanded
as they are local to the chroot environment.

For example, systemctl creates absolute symlinks to enable new services
below the directory /etc/systemd.

The variable FAKECHROOT_EXCLUDED_PATH was introduced to disable the
expansion for the canonical file names with an absolute path that are
listed; but it do not applies for the value of the symbolic name.

This introduces the variable FAKECHROOT_EXCLUDE_SYMLINK to disable the
expansion for the symlink values with an absolute path that are listed.