New variable FAKECHROOT_INCLUDE_PATH is introduced
This new variable FAKECHROOT_INCLUDE_PATH is analogous to the variable
FAKECHROOT_EXCLUDE_PATH but do the opposite. The paths that are listed
in that variable *AND* they are absolute paths are assumed local to the
chroot'ed environment are expanded with the chroot base directory.
FAKECHROOT_INCLUDE_PATH takes precendence over FAKECHROOT_EXCLUDE_PATH
for the local evaluation.