Add the replacement executable for ld-musl-x86_64.so.1
This replacement can be used to substitute a command in the chroot'ed
environment in the libc differs with the one in the host. It uses the
musl loader available on the host to run the command in the chroot'ed
environment.

In the end, this allows to create a musl rootfs within a glibc's host
system with the same architecature.

For example, the alpine-make-rootfs environment, used to make a musl's
rootfs, substitutes the /bin/busybox commands that are linked to the
musl library in the chroot environment with the ld-musl-x86_64.so.1
replacent that runs the musl loader installed in the glib's host system
to run it. Otherwise, it is glibc's loader ld-linux-x86-64.so.2 that is
used and fails run /bin/busybox because of missing glibc's specific
symbols.

Note: The option --argv0 /bin/busybox sets the symlink to /bin/busybox
instead of its full path (i.e. /path/to/chrooted/dir/bin/busybox).

Note: The alpine-make-rootfs environment runs the shell interpreter from
the host by excluding. The symlink is local to the chroot'ed environment
and thus is it broken if busybox is not available to the host.