New variable FAKECHROOT_EXCLUDE_SYMLINK_STRING is introduced
This new variable FAKECHROOT_EXCLUDE_SYMLINK_STRING is analogous to the
variable FAKECHROOT_EXCLUDE_SYMLINK but the exclusion is based on the
value instead of the path. The symlink values that are absolute paths
*AND* that are created by either symlink(3p) or symlinkat(3p) *AND* that
are listed in the variable FAKECHROOT_EXCLUDE_SYMLINK_STRING are *NOT*
expanded with the chroot base directory: the value remain local to the
chroot. However, the canonical file name path is still expanded.
FAKECHROOT_EXCLUDE_SYMLINK_STRING takes precendence over
FAKECHROOT_EXCLUDE_SYMLINK for the local evaluation.